backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.0
yfinance==0.2.36
ta==0.11.0
requests==2.31.0
pytest==8.0.0
httpx==0.26.0 